requirements.txt
requests
sendgrid 
pandas
python-dotenv
plotly